+++ spigot/src/main/java/one/oth3r/directionhud/utils/BossBarManager.java
@@ -29,9 +29,9 @@ public class BossBarManager {
         bossBar.setColor(BarColor.valueOf(((String) player.getPData().getHud().getSetting(HUD.Setting.bossbar__color)).toUpperCase()));
         if (Destination.dest.get(player).hasXYZ() && (boolean) player.getPData().getHud().getSetting(HUD.Setting.bossbar__distance)) {
             int dist = Destination.dest.getDist(player);
-            double progress = getProgress(dist,(double) player.getPData().getHud().getSetting(HUD.Setting.bossbar__distance_max));
+            double progress = getProgress(dist,(int) player.getPData().getHud().getSetting(HUD.Setting.bossbar__distance_max));
             bossBar.setStyle(BarStyle.SEGMENTED_10);
-            if ((double) player.getPData().getHud().getSetting(HUD.Setting.bossbar__distance_max) == 0) {
+            if ((int) player.getPData().getHud().getSetting(HUD.Setting.bossbar__distance_max) == 0) {
                 progress = getProgress(dist,1000);
                 StringBuilder s = new StringBuilder();
                 for (int i = 1;i<5;i++) {
